+++ MarkUS-Backend/src/database/migrations/20210723214455-create-evaluable.js
@@ -40,7 +40,7 @@ module.exports = {
         },
         onDelete: 'cascade'
       },
-      mark: Sequelize.INTEGER,
+      mark: Sequelize.DOUBLE,
       weight: Sequelize.INTEGER,  
         createdAt: {
         allowNull: false,
